Final delete drawings script

+++ main.py
@@ -196,17 +196,18 @@ def delete_old_drawings():
         project_id_query = 'SELECT project_id from projects WHERE project_number='+project_number
         cur.execute(project_id_query)
         res = cur.fetchone()
-        project_id = res[0]
-        get_drawings_for_projects = 'SELECT pdf FROM Docs WHERE project_id='+str(project_id)+' AND folder!="RECEIPTS" AND folder!="AGREEMENT "'
-        cur.execute(get_drawings_for_projects)
-        res = cur.fetchall()
-        for d in res:
-            try:
-                os.remove('/home/buildahome2016/public_html/app.buildahome.in/team/Drawings/'+d[0])
-            except:
-                pass
-        delete_drawing_query = 'DELETE from Docs WHERE project_id='+str(project_id)+' AND folder!="RECEIPTS" AND folder!="AGREEMENT "'
-        cur.execute(delete_drawing_query)
+        if res is not None:
+            project_id = res[0]
+            get_drawings_for_projects = 'SELECT pdf FROM Docs WHERE project_id='+str(project_id)+' AND folder!="RECEIPTS" AND folder!="AGREEMENT "'
+            cur.execute(get_drawings_for_projects)
+            res = cur.fetchall()
+            for d in res:
+                try:
+                    os.remove('/home/buildahome2016/public_html/app.buildahome.in/team/Drawings/'+d[0])
+                except:
+                    pass
+            delete_drawing_query = 'DELETE from Docs WHERE project_id='+str(project_id)+' AND folder!="RECEIPTS" AND folder!="AGREEMENT "'
+            cur.execute(delete_drawing_query)
